계정 삭제 기능 예외 처리

+++ app/src/main/java/nepal/swopnasansar/AccountantMainActivity.kt
@@ -2,7 +2,6 @@ package nepal.swopnasansar
 
 import android.content.Intent
 import android.os.Bundle
-import android.util.Log
 import android.view.Menu
 import android.view.MenuItem
 import android.view.View
@@ -10,7 +9,12 @@ import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.lifecycle.lifecycleScope
+import com.google.firebase.FirebaseException
+import com.google.firebase.FirebaseNetworkException
 import com.google.firebase.auth.EmailAuthProvider
+import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException
+import com.google.firebase.auth.FirebaseAuthInvalidUserException
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
@@ -78,35 +82,70 @@ class AccountantMainActivity: AppCompatActivity() {
     override fun onOptionsItemSelected(item: MenuItem): Boolean {
         when(item.itemId) {
             R.id.item_sign_out -> {
-                val dialog = SignoutDialog(this)
-                dialog.setOnClickListener(object: SignoutDialog.ButtonClickListener {
+                val dialog = SignOutDialog(this)
+                dialog.setOnClickListener(object: SignOutDialog.ButtonClickListener {
                     override fun onClicked(email: String, pw: String) {
-                        val user = authDao.getUser()
+                        if (email.isNotBlank() || pw.isNotBlank())  {
+                            val user = authDao.getUser()
 
-                        if (user != null) {
-                            val credential = EmailAuthProvider.getCredential(email, pw)
-                            user.reauthenticate(credential)
-                                .addOnSuccessListener {
-                                    user.delete()
-
+                            if (user != null) {
+                                try {
                                     lifecycleScope.launch {
-                                        val deleteDBResult = withContext(Dispatchers.IO) {
-                                            accountantDao.removeAccountantByKey(uid!!)
-                                        }
+                                        binding.pbAccountantMain.visibility = View.VISIBLE
 
-                                        if (deleteDBResult) {
-                                            Toast.makeText(applicationContext, "Your account has been deleted.", Toast.LENGTH_LONG).show()
-                                            val intent = Intent(this@AccountantMainActivity, CheckRoleActivity::class.java)
-                                            startActivity(intent)
-                                        }
+                                        val credential = EmailAuthProvider.getCredential(email, pw)
+                                        user.reauthenticate(credential)
+                                            .addOnSuccessListener {
+                                                CoroutineScope(Dispatchers.IO).launch {
+                                                    val deleteDBResult = withContext(Dispatchers.IO) {
+                                                        accountantDao.removeAccountantByKey(uid!!)
+                                                    }
+
+                                                    withContext(Dispatchers.Main) {
+                                                        if (deleteDBResult) {
+                                                            user.delete()
+                                                                .addOnSuccessListener {
+                                                                    Toast.makeText(applicationContext, "Your account has been deleted.", Toast.LENGTH_LONG).show()
+                                                                    binding.pbAccountantMain.visibility = View.GONE
+                                                                    val intent = Intent(this@AccountantMainActivity, CheckRoleActivity::class.java)
+                                                                    startActivity(intent)
+                                                                }
+                                                                .addOnFailureListener {
+                                                                    Toast.makeText(this@AccountantMainActivity, "An error has occurred. Please contact administrator", Toast.LENGTH_LONG).show()
+                                                                    binding.pbAccountantMain.visibility = View.GONE
+                                                                }
+                                                        } else {
+                                                            Toast.makeText(this@AccountantMainActivity, "An error has occurred. Please contact administrator.", Toast.LENGTH_SHORT).show()
+                                                            binding.pbAccountantMain.visibility = View.GONE
+                                                        }
+                                                    }
+                                                }
+
+                                            }
+                                            .addOnFailureListener {
+                                                Toast.makeText(this@AccountantMainActivity, "Email or password is incorrect. Please check again.", Toast.LENGTH_LONG).show()
+                                                binding.pbAccountantMain.visibility = View.GONE
+                                            }
+
+
                                     }
+                                } catch (e: FirebaseAuthInvalidCredentialsException) {
+                                    Toast.makeText(this@AccountantMainActivity, "Email or password is incorrect. Please check again.", Toast.LENGTH_LONG).show()
+                                    binding.pbAccountantMain.visibility = View.GONE
+                                } catch (e: FirebaseNetworkException) {
+                                    Toast.makeText(this@AccountantMainActivity, "Network issue has occurred. Please try again later.", Toast.LENGTH_LONG).show()
+                                    binding.pbAccountantMain.visibility = View.GONE
+                                } catch (e: FirebaseException) {
+                                    Toast.makeText(this@AccountantMainActivity, "A DB service error has occurred. Please try again later.", Toast.LENGTH_LONG).show()
+                                    binding.pbAccountantMain.visibility = View.GONE
+                                } catch (e: FirebaseAuthInvalidUserException) {
+                                    Toast.makeText(this@AccountantMainActivity, "User does not exist.", Toast.LENGTH_LONG).show()
+                                    binding.pbAccountantMain.visibility = View.GONE
+                                } catch (e: Exception) {
+                                    Toast.makeText(this@AccountantMainActivity, "An error has occurred. Please try again later.", Toast.LENGTH_LONG).show()
+                                    binding.pbAccountantMain.visibility = View.GONE
                                 }
-                                .addOnFailureListener {
-                                    Toast.makeText(this@AccountantMainActivity, "Fail to delete your account. Try again.", Toast.LENGTH_SHORT).show()
-                                }
-                                .addOnFailureListener {
-                                    Toast.makeText(this@AccountantMainActivity, "Authentication has failed.", Toast.LENGTH_SHORT).show()
-                                }
+                            }
                         }
                     }
                 })
@@ -142,5 +181,4 @@ class AccountantMainActivity: AppCompatActivity() {
     override fun onBackPressed() {
         // 뒤로 가기 버튼 동작 없음
     }
-
 }
